praying god this is the last api commit

+++ backend/src/main/java/com/controler/Transactions.java
@@ -2,7 +2,8 @@ package com.controler;
 
 
 
-import com.classes.Transactions;
+import com.classes.*;
+import com.repos.BankRepository;
 import com.repos.TransactionsRepository;
 
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -27,6 +28,9 @@ class transactionsControler{
 
     @Autowired
     TransactionsRepository transactionsRepository;
+
+	@Autowired
+    BankRepository bankRepository;
 
 	@GetMapping()
 	public String teste(){
@@ -59,32 +63,75 @@ class transactionsControler{
 	@PostMapping("/addTransactions")
 	public ResponseEntity<Transactions> addTransactions(@RequestBody Transactions transactions) {
 		try {
-            Transactions transactionsObj = transactionsRepository.save(transactions);
-            return new ResponseEntity<>(transactionsObj, HttpStatus.CREATED);
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-	}
-	@PostMapping("/updateTransactions/{id}")
-	public ResponseEntity<Transactions> updateTransactions(@RequestBody Transactions transactions,@PathVariable Long id) {
-		Optional<Transactions> oldTransactionsData = transactionsRepository.findById(id);
-		if (oldTransactionsData.isPresent()){
-			Transactions updateTransactions = oldTransactionsData.get();
-			updateTransactions.setMoneyBank(transactions.getMoneyBank());
-			// updateTransactions.setTransactionsList(transactions.getTransactionsList());
-			Transactions transactionsData = transactionsRepository.save(updateTransactions);
-			return new ResponseEntity<>(transactionsData,HttpStatus.OK);
-		}else{
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			Transactions transactionsObj = transactionsRepository.save(transactions);
+			List<Bank> bankObjList = bankRepository.findByBankName(transactions.getBankName());
+			if (!bankObjList.isEmpty()) {
+				for (Bank bank : bankObjList) {
+					bank.setMoneyBank(bank.getMoneyBank() + transactions.getMoneyBank());
+					bankRepository.save(bank); 
+				}
+			} else {
+				
+				Bank newBank = new Bank();
+				newBank.setBankName(transactions.getBankName());
+				newBank.setBankPhoto(transactions.getBankPhoto());
+				newBank.setClientId(transactions.getClientId());
+				newBank.setMoneyBank(transactions.getMoneyBank());
+				bankRepository.save(newBank);
+			}
+			return new ResponseEntity<>(transactionsObj, HttpStatus.CREATED);
+		} catch (Exception e) {
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
 		}
-		
 	}
 
-	@DeleteMapping("/deleteById/{id}")
-	public ResponseEntity<Transactions> deleteById(@PathVariable Long id){
-		transactionsRepository.deleteById(id);
-		return new ResponseEntity<>(HttpStatus.OK);
+	@PostMapping("/updateTransactions/{id}")
+	public ResponseEntity<Transactions> updateTransactions(@RequestBody Transactions transactions, @PathVariable Long id) {
+		Optional<Transactions> oldTransactionsData = transactionsRepository.findById(id);
+		if (oldTransactionsData.isPresent()) {
+			Transactions oldTransaction = oldTransactionsData.get();
+
+			float amountDifference = transactions.getMoneyBank() - oldTransaction.getMoneyBank();
+
+			oldTransaction.setMoneyBank(transactions.getMoneyBank());
+
+			Transactions updatedTransaction = transactionsRepository.save(oldTransaction);
+
+			List<Bank> bankObjList = bankRepository.findByBankName(oldTransaction.getBankName());
+			if (!bankObjList.isEmpty()) {
+				for (Bank bank : bankObjList) {
+					bank.setMoneyBank(bank.getMoneyBank() + amountDifference);
+					bankRepository.save(bank); 
+				}
+			}
+
+			return new ResponseEntity<>(updatedTransaction, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 	}
+
+
+	@DeleteMapping("/deleteById/{id}")
+public ResponseEntity<Void> deleteById(@PathVariable Long id) {
+    Optional<Transactions> transactionData = transactionsRepository.findById(id);
+    if (transactionData.isPresent()) {
+        Transactions transaction = transactionData.get();
+
+        List<Bank> bankObjList = bankRepository.findByBankName(transaction.getBankName());
+        if (!bankObjList.isEmpty()) {
+            for (Bank bank : bankObjList) {
+                bank.setMoneyBank(bank.getMoneyBank() - transaction.getMoneyBank());
+                bankRepository.save(bank); 
+            }
+        }
+        transactionsRepository.deleteById(id);
+        return new ResponseEntity<>(HttpStatus.OK);
+    } else {
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+}
+
 	
 
 }
